tools/requirements.txt
# Image Generation Requirements
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
